refactor: move existing tests into docker_integration test module
